Move integration test helpers to dcoscli
This makes them easily usable from other packages.

This would be useful for DCOS_OSS-1756 (#1110).
