+++ cli/dcoscli/test/common.py
