+++ cli/dcoscli/test/marathon.py
