+++ cli/dcoscli/test/package.py
