+++ cli/dcoscli/test/service.py
