+++ cli/tests/integrations/test_cluster_migration.py
@@ -6,7 +6,7 @@ import pytest
 
 from dcos import config, constants, util
 
-from .helpers.common import assert_command, exec_command
+from dcoscli.test.common import assert_command, exec_command
 
 
 @pytest.fixture(scope="module")
